# contracts/openapi/suilearn-v2.yaml
openapi: 3.0.3
info:
  title: SuiLearn V2 API
  version: 0.1.0
  description: API contract for Java backend AI/RAG and the React Web knowledge base workbench.
servers:
  - url: http://localhost:8080
paths:
  /api/v2/knowledge-bases:
    get:
      summary: List knowledge bases
      responses:
        "200":
          description: Knowledge base list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeBase"
    post:
      summary: Create a knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "200":
          description: Created knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
  /api/v2/knowledge-bases/{knowledgeBaseId}:
    get:
      summary: Get knowledge base detail
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Knowledge base detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseDetail"
    patch:
      summary: Rename a knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameKnowledgeBaseRequest"
      responses:
        "200":
          description: Updated knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    delete:
      summary: Delete a knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "204":
          description: Deleted
  /api/v2/knowledge-bases/{knowledgeBaseId}/questions:
    get:
      summary: List questions saved in a knowledge base
      description: Lightweight workbench view for knowledge base detail. Full Web practice flow belongs to the next phase.
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Question summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionSummary"
  /api/v2/knowledge-bases/{knowledgeBaseId}/statistics:
    get:
      summary: Get lightweight learning statistics for a knowledge base
      description: Supports the second-version Web workbench detail page without introducing the full Web learning flow.
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Knowledge base statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseStatistics"
  /api/v2/knowledge-bases/{knowledgeBaseId}/materials:
    get:
      summary: List materials in a knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Material list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningMaterial"
    post:
      summary: Import Markdown, TXT, or PDF material
      description: >
        Imports material into the knowledge base. MVP implementations accept normalized text content plus
        title, fileName, and sourceType. multipart/form-data is reserved as the cross-client upload shape,
        but the MVP must not claim that raw PDF binary parsing is available; PDF binary parsing is a later
        document-parser adapter concern. When sourceType is PDF during MVP, content is the already textified
        PDF content supplied by the client or another upstream extractor.
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportMaterialRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadMaterialFormData"
      responses:
        "200":
          description: Imported material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningMaterial"
  /api/v2/materials/{materialId}:
    get:
      summary: Get material detail
      parameters:
        - $ref: "#/components/parameters/MaterialId"
      responses:
        "200":
          description: Material detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialDetail"
    delete:
      summary: Delete material
      description: Soft deletes the material, removes it from future RAG/search/generation, invalidates chunks and embeddings, deletes pending generated content by default, and keeps already saved questions or AI notes with deleted-source SourceRef metadata.
      parameters:
        - $ref: "#/components/parameters/MaterialId"
        - in: query
          name: savedContentPolicy
          schema:
            $ref: "#/components/schemas/DeletedMaterialSavedContentPolicy"
          description: Policy for questions or AI notes already saved from this material. Default is KEEP_SAVED_CONTENT.
        - in: query
          name: pendingContentPolicy
          schema:
            $ref: "#/components/schemas/DeletedMaterialPendingContentPolicy"
          description: Policy for pending generated drafts sourced from this material. Default is DELETE_PENDING_GENERATED_CONTENT.
      responses:
        "200":
          description: Material deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialDeletionResult"
  /api/v2/materials/{materialId}/extract-knowledge-points:
    post:
      summary: Extract editable knowledge points from material
      description: >
        Creates a persisted generation task even when the MVP executes synchronously.
        The returned task id can be queried through /api/v2/tasks/{taskId}.
      parameters:
        - $ref: "#/components/parameters/MaterialId"
      responses:
        "200":
          description: Extracted knowledge points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgePointExtractionResult"
  /api/v2/knowledge-bases/{knowledgeBaseId}/knowledge-points:
    get:
      summary: List extracted or generated knowledge points
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Knowledge point list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgePoint"
  /api/v2/knowledge-points/{knowledgePointId}:
    patch:
      summary: Edit extracted knowledge point
      parameters:
        - $ref: "#/components/parameters/KnowledgePointId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKnowledgePointRequest"
      responses:
        "200":
          description: Updated knowledge point
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgePoint"
    delete:
      summary: Delete extracted knowledge point
      parameters:
        - $ref: "#/components/parameters/KnowledgePointId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    KnowledgeBaseId:
      in: path
      name: knowledgeBaseId
      required: true
      schema:
        type: string
    MaterialId:
      in: path
      name: materialId
      required: true
      schema:
        type: string
...
    KnowledgePointId:
      in: path
      name: knowledgePointId
      required: true
      schema:
        type: string
...
  schemas:
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    RenameKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    KnowledgeBase:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KnowledgeBaseDetail:
      allOf:
        - $ref: "#/components/schemas/KnowledgeBase"
        - type: object
          description: >
            Lightweight detail counters. Counts are computed from persisted rows in the
            knowledge base and must not be placeholder constants.
          properties:
            materialCount:
              type: integer
              description: Non-deleted materials in this knowledge base.
            knowledgePointCount:
              type: integer
              description: Non-deleted extracted or generated knowledge points in this knowledge base.
            questionCount:
              type: integer
              description: Saved formal questions in this knowledge base, including saved AI-generated questions.
            generatedContentCount:
              type: integer
              description: Generated contents not soft-deleted, including pending review, saved, and discarded items.
            aiNoteCount:
              type: integer
              description: Saved AI notes in this knowledge base.
    QuestionSummary:
      type: object
      required: [id, knowledgeBaseId, questionType, stem, categoryId, categoryName, knowledgePointIds, createdAt, sourceRefs]
      description: Saved formal question summary. Saved AI-generated questions must retain formal category and knowledge point attribution from the reviewed draft.
      properties:
        id:
          type: string
        knowledgeBaseId:
          type: string
        questionType:
          $ref: "#/components/schemas/QuestionType"
        stem:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        knowledgePointIds:
          type: array
          items:
            type: string
        answeredCount:
          type: integer
        correctRate:
          type: number
          format: double
        sourceRefs:
          type: array
          items:
            $ref: "#/components/schemas/SourceRef"
        createdAt:
          type: string
          format: date-time
        savedAt:
          type: string
          format: date-time
    KnowledgeBaseStatistics:
      type: object
      required: [knowledgeBaseId, questionCount, answeredQuestionCount, answerCount, wrongQuestionCount]
      description: >
        Lightweight learning statistics for the knowledge base detail page. Values must be
        derived from saved questions and answer records; if no answer records exist, counts
        are zero and correctRate is omitted or null rather than filled with a placeholder.
      properties:
        knowledgeBaseId:
          type: string
        questionCount:
          type: integer
          description: Saved formal questions in the knowledge base.
        answeredQuestionCount:
          type: integer
          description: Distinct saved questions with at least one answer record.
        answerCount:
          type: integer
          description: Total answer attempts recorded for saved questions in the knowledge base.
        correctRate:
          type: number
          format: double
          nullable: true
          minimum: 0
          maximum: 1
          description: Correct answers divided by answerCount; null or omitted when answerCount is 0.
        wrongQuestionCount:
          type: integer
          description: Saved questions that currently have at least one unmastered wrong-question record.
        weakKnowledgePointIds:
          type: array
          description: Knowledge point ids ranked as weak by wrong-question or low-correct-rate evidence; empty when there is no evidence.
          items:
            type: string
    ImportMaterialRequest:
      type: object
      required: [title, sourceType, content]
      description: >
        JSON import shape for MVP. content is normalized plain text or Markdown content, even when
        sourceType is PDF. Raw PDF binary parsing is intentionally not promised by this contract version.
      properties:
        title:
          type: string
        fileName:
          type: string
          description: Original client-side file name when the material came from a file.
        sourceType:
          $ref: "#/components/schemas/MaterialSourceType"
        content:
          type: string
          description: Textified material content used for chunking, embedding, search, and RAG in the MVP.
    UploadMaterialFormData:
      type: object
      required: [title, sourceType, content]
      description: >
        multipart/form-data shape kept for Web/Android upload compatibility. The MVP server consumes
        title, sourceType, fileName, and textified content. The optional file part may be sent for future
        compatibility, but raw PDF binary parsing is not part of the MVP behavior until a parser adapter
        is wired in.
      properties:
        title:
          type: string
        fileName:
          type: string
          description: Original file name to display and keep for traceability.
        sourceType:
          $ref: "#/components/schemas/MaterialSourceType"
        content:
          type: string
          description: Textified Markdown, TXT, or PDF content consumed by the MVP.
        file:
          type: string
          format: binary
          description: Optional original file part retained for client compatibility; MVP behavior must not depend on parsing PDF binary content.
    LearningMaterial:
      type: object
      required: [id, knowledgeBaseId, title, sourceType, status, importTaskId, createdAt]
      description: Material list item. It intentionally excludes original content; call material detail for content or chunk data.
      properties:
        id:
          type: string
        knowledgeBaseId:
          type: string
        title:
          type: string
        sourceType:
          $ref: "#/components/schemas/MaterialSourceType"
        status:
          $ref: "#/components/schemas/MaterialStatus"
        importTaskId:
          type: string
          description: Material import task id. Present even when import completed synchronously.
        embeddingTaskId:
          type: string
          description: Embedding task id once chunk embedding has started or completed.
        errorMessage:
          type: string
          description: Short user-safe failure summary when status is FAILED.
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    MaterialDetail:
      allOf:
        - $ref: "#/components/schemas/LearningMaterial"
        - type: object
          description: Material detail may include the normalized source content used by the MVP and derived chunks.
          properties:
            content:
              type: string
              description: Full normalized text content when the caller is allowed to inspect the material body.
            contentPreview:
              type: string
              description: Short preview for clients that do not need or cannot fetch full content.
            chunks:
              type: array
              items:
                $ref: "#/components/schemas/MaterialChunk"
            extractedKnowledgePoints:
              type: array
              items:
                $ref: "#/components/schemas/KnowledgePoint"
    MaterialChunk:
      type: object
      required: [id, knowledgeBaseId, materialId, content, ordinal, embeddingStatus]
      description: >
        Smallest retrievable unit for semantic search and RAG. The embedding vector itself
        is stored server-side in pgvector and is not returned by the API.
      properties:
        id:
          type: string
        knowledgeBaseId:
          type: string
        materialId:
          type: string
        content:
          type: string
        ordinal:
          type: integer
        titlePath:
          type: array
          items:
            type: string
          description: Heading path or structural labels used to help users locate the chunk.
        tokenEstimate:
          type: integer
          description: Approximate token count used by retrieval and prompt budgeting.
        embeddingStatus:
          $ref: "#/components/schemas/EmbeddingStatus"
        embeddingModel:
          type: string
          description: Embedding model used for the stored pgvector value; never an API key or secret.
        embeddingDimensions:
          type: integer
          description: Dimension count of the stored vector, used for migration and compatibility checks.
        sourceRef:
          $ref: "#/components/schemas/SourceRef"
    MaterialDeletionResult:
      type: object
      required: [materialId, status, savedContentPolicy, pendingContentPolicy, invalidatedChunkCount, deletedPendingGeneratedContentCount]
      description: >
        Result of soft-deleting a material. Deleted materials are excluded from future search, RAG, and
        generation. Pending generated drafts sourced from the material are deleted by default. Already
        saved questions and AI notes are kept by default, with their SourceRef.deleted flag set to true.
      properties:
        materialId:
          type: string
        status:
          $ref: "#/components/schemas/MaterialStatus"
        savedContentPolicy:
          $ref: "#/components/schemas/DeletedMaterialSavedContentPolicy"
        pendingContentPolicy:
          $ref: "#/components/schemas/DeletedMaterialPendingContentPolicy"
        invalidatedChunkCount:
          type: integer
        deletedPendingGeneratedContentCount:
          type: integer
        retainedSavedQuestionCount:
          type: integer
        retainedAiNoteCount:
          type: integer
        deletedAt:
          type: string
          format: date-time
    KnowledgePoint:
      type: object
      required: [id, knowledgeBaseId, name, description]
      properties:
        id:
          type: string
        knowledgeBaseId:
          type: string
        name:
          type: string
        description:
          type: string
        sourceMaterialId:
          type: string
          deprecated: true
        sourceRefs:
          type: array
          items:
            $ref: "#/components/schemas/SourceRef"
    KnowledgePointExtractionResult:
      type: object
      required: [taskId, task, knowledgePoints]
      properties:
        taskId:
          type: string
        task:
          $ref: "#/components/schemas/TaskStatus"
        knowledgePoints:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgePoint"
    UpdateKnowledgePointRequest:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
...
    MaterialSourceType:
      type: string
      enum: [MARKDOWN, TXT, PDF]
    MaterialStatus:
      type: string
      enum: [UPLOADED, PARSING, CHUNKING, INDEXING, READY, FAILED, DELETED]
    EmbeddingStatus:
      type: string
      enum: [PENDING, INDEXING, READY, FAILED, INVALIDATED]
      description: >
        Server-side embedding lifecycle for material chunks. INVALIDATED means the source
        material was deleted or the embedding model changed and the vector must not be used.
    DeletedMaterialSavedContentPolicy:
      type: string
      enum: [KEEP_SAVED_CONTENT, DELETE_SAVED_CONTENT]
      default: KEEP_SAVED_CONTENT
      description: >
        KEEP_SAVED_CONTENT retains already saved questions and AI notes, marking their material SourceRef
        objects as deleted. DELETE_SAVED_CONTENT removes saved content derived solely from the deleted
        material; mixed-source content should be retained with only the deleted source marked.
    DeletedMaterialPendingContentPolicy:
      type: string
      enum: [DELETE_PENDING_GENERATED_CONTENT, KEEP_PENDING_GENERATED_CONTENT]
      default: DELETE_PENDING_GENERATED_CONTENT
      description: >
        DELETE_PENDING_GENERATED_CONTENT marks pending generated drafts sourced from the deleted material
        as DELETED. KEEP_PENDING_GENERATED_CONTENT keeps drafts reviewable, but their material SourceRef
        objects must have deleted=true.
...
    AiProviderType:
      type: string
      enum: [FAKE, OPENAI_COMPATIBLE]
    SourceType:
      type: string
      enum: [KNOWLEDGE_POINT, WRONG_QUESTION, MATERIAL, MATERIAL_CHUNK, KNOWLEDGE_BASE, ANSWER_HISTORY]
    SourceRef:
      type: object
      required: [type, id]
      description: >
        Traceability pointer for generated and saved content. If the referenced material or chunk is
        soft-deleted after content is saved, keep the SourceRef for auditability and set deleted=true.
      properties:
        type:
          $ref: "#/components/schemas/SourceType"
        id:
          type: string
        knowledgeBaseId:
          type: string
        title:
          type: string
        materialId:
          type: string
        chunkId:
          type: string
        deleted:
          type: boolean
          description: True when the referenced source, such as a material or material chunk, has been soft deleted and must no longer be used for new search, RAG, or generation.
        excerpt:
          type: string
    QuestionType:
      type: string
      enum: [SINGLE_CHOICE, MULTIPLE_CHOICE, TRUE_FALSE, SHORT_ANSWER]
...
    TaskStatus:
      type: object
      required: [id, kind, status, createdAt, updatedAt]
      description: >
        Persisted status for material import, embedding, and AI generation work. Synchronous
        MVP execution still creates a task and finishes it before returning from the initiating API.
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/TaskKind"
        status:
          $ref: "#/components/schemas/TaskLifecycleStatus"
        knowledgeBaseId:
          type: string
        materialId:
          type: string
        generatedContentId:
          type: string
        providerType:
          $ref: "#/components/schemas/AiProviderType"
        model:
          type: string
          description: Chat or embedding model used by this task; never contains secrets.
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
        currentStep:
          type: string
          description: User-safe step label such as PARSING, CHUNKING, EMBEDDING, GENERATING.
        errorCode:
          type: string
        errorMessage:
          type: string
          description: User-safe error summary. Do not store or return provider raw secrets.
        retryCount:
          type: integer
        resultRef:
          $ref: "#/components/schemas/TaskResultRef"
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskResultRef:
      type: object
      properties:
        type:
          type: string
          enum: [MATERIAL, MATERIAL_CHUNKS, GENERATED_CONTENT, AI_NOTE_DRAFT, KNOWLEDGE_POINTS]
        id:
          type: string
        count:
          type: integer
    TaskKind:
      type: string
      enum: [MATERIAL_IMPORT, EMBEDDING, QUESTION_GENERATION, KNOWLEDGE_POINT_EXTRACTION, EXPLANATION_GENERATION, REVIEW_SUGGESTION_GENERATION]
    TaskLifecycleStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]